pyproject.toml
[tool.poetry]
name = "market-test"
version = "0.1.0"
description = "test job for market company"
authors = ["sizlik <zenjagems@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
daphne = "^4.1.0"
djangorestframework = "^3.15.1"
channels-redis = "^4.2.0"
channels = {extras = ["daphne"], version = "^4.1.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
